Moved app contract query handlers
