+++ contracts/app/andromeda-app-contract/src/lib.rs
@@ -8,3 +8,4 @@ pub mod mock;
 pub mod testing;
 
 mod execute;
+mod query;
